add upscale, refactor, change name
add upscale, refactor, change name
