+++ .gitignore
@@ -1,3 +1,3 @@
-RifeTensorRT/out/*
+FrameSmith/out/*
 .vs/*
