+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "RifeTensorRT/concurrentqueue"]
+	path = RifeTensorRT/concurrentqueue
+	url = https://github.com/cameron314/concurrentqueue.git
